# pluggables-app/target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Pluggables app REST API","description":"Pluggables application REST API"},"basePath":"/onos/pluggables-app","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"pluggables","description":"Query Pluggables app"}],"paths":{"/pluggables/configure":{"post":{"tags":["pluggables"],"summary":"Set a Pluggables configuration and create a connection","description":"","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/ConfpluggablesPost"},"description":"input JSON","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/pluggables/delete/{connection-name}":{"delete":{"tags":["pluggables"],"summary":"Disable the pluggables tha belonging to a certain connection","description":"","parameters":[{"name":"connection-name","in":"path","type":"string","description":"connection identifier","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/pluggables":{"get":{"tags":["pluggables"],"summary":"Gets all Pluggables","description":"Returns an array of all Pluggables in the network.","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{"ConfpluggablesPost":{"type":"object","title":"confdevicepluggables","required":["pluggable-one","pluggable-two","srcConnectPoint","dstConnectPoint"],"properties":{"pluggable-one":{"type":"object","title":"pluggable-one","required":["device-id","pluggable-id","opmode","frequency","power"],"properties":{"device-id":{"type":"String","example":"device:10.30.2.44:50001"},"pluggable-id":{"type":"String","example":"Cisco-400zr-QSFP-DD-01"},"opmode":{"type":"String","example":"1"},"frequency":{"type":"Double","example":192300.0},"power":{"type":"Float","example":1.0}}},"pluggable-two":{"type":"object","title":"pluggable-two","required":["device-id","pluggable-id","opmode","frequency","power"],"properties":{"device-id":{"type":"String","example":"device:10.30.2.102:50001"},"pluggable-id":{"type":"String","example":"Cisco-400zr-QSFP-DD-01"},"opmode":{"type":"String","example":"1"},"frequency":{"type":"Double","example":192300.0},"power":{"type":"Float","example":1.0}}},"srcConnectPoint":{"type":"String","example":"device:10.30.2.46:50001/1"},"dstConnectPoint":{"type":"String","example":"device:10.30.2.102:50001/1"},"connection-name":{"type":"String","example":"connection-1"}}}}}